+++ 3dgs_renderer_v3.py
@@ -69,7 +69,7 @@ def rasterize_tiles(
     # which the builder has already sorted near to far. Finish with the background
     # weighted by the remaining transmittance, matching 3dgs_renderer_v1.
     color = wp.vec3(0.0, 0.0, 0.0)
-    transmittance = 1.0
+    transmittance = wp.float32(1.0)
     
     tile_start = tile_offsets[tile]
     tile_end = tile_offsets[tile + 1]
